drawing elements in the image preview
