+++ src/mainmenu/blades/imageEdit/EditImageBlade.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+import { Component } from "../../../CarbonFlux";
+import { BladeBody } from "../BladePage";
+import ImageEditor from "./ImageEditor";
+
+interface IEditImageBladeProps {
+    image: string;
+
+    onComplete: (image: string) => void;
+}
+
+export default class EditImageBlade extends Component<IEditImageBladeProps> {
+    render() {
+        return <BladeBody>
+            <ImageEditor image={this.props.image} onComplete={this.props.onComplete}/>
+        </BladeBody>
+    }
+
+    static contextTypes = {
+        intl: React.PropTypes.object,
+        currentBladeId: React.PropTypes.number,
+        bladeContainer: React.PropTypes.any
+    }
+}
