+++ src/styles/less/__main-menu.less
@@ -474,7 +474,6 @@
 
 
     &__crop-area {
-        .befratio;
         .rel;
         .clip;
         .wbg(3);
