+++ src/styles/less/_blades.less
@@ -206,22 +206,7 @@
     margin-top: @line_air / 2;
 
     &-item {
-        width: 100%;
-        .point;
-        padding: .5em .5em;
-        margin-bottom: 1px;
 
-        .wbg(3);
-
-        &:hover {
-            .wbg(6);
-        }
-
-        &_selected {
-            &, &:hover {
-                .wbg(10);
-            }
-        }
 
     }
 }
