+++ src/styles/less/_ui.less
@@ -89,6 +89,7 @@
 
             .rel;
             .h100;
+            .w100;
             display: inline-flex;
             align-items: center;
             flex-wrap: nowrap;
@@ -102,9 +103,24 @@
             //.middler();
             //& > * { .inbl_middle; }
             //.inbl;
-            width: 100%;
             .nowrap;
             .tleft;
+
+            .point;
+            padding: .5em .5em;
+            margin-bottom: 1px;
+
+            .wbg(3);
+
+            &:hover {
+                .wbg(6);
+            }
+
+            &_selected {
+                &, &:hover {
+                    .wbg(10);
+                }
+            }
 
 
             & > .ico-triangle {
